chore: Move regosig/ back to tracee-rules.
