+++ tracee-rules/regosig/mapper.go
