+++ tracee-rules/regosig/mapper_test.go
