+++ tracee-rules/regosig/traceerego.go
